+++ changelog.md
@@ -22,6 +22,30 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ---
 
+## [v0.3.0] - 2023/07/30
+
+### Added
+
+- Jest test suite (thanks @velebit!)
+- Setting to preserve leading characters if part of an indented list
+- Setting to control whether `CTRL/CMD + S` trims document
+
+### Changed
+
+- Auto-Trim to skip currently-active text block ([\#5](https://github.com/zlovatt/obsidian-trim-whitespace/issues/5))
+- Trimming via `CTRL/CMD + S` no longer requires Auto-Trim to be enabled ([\#12](https://github.com/zlovatt/obsidian-trim-whitespace/issues/12))
+- "Skip Code Blocks" setting name to "Preserve Code Blocks"
+- Improved detection of code blocks
+
+### Fixed
+
+- Trimming of indented lists ([\#9](https://github.com/zlovatt/obsidian-trim-whitespace/issues/9))
+- Plugin not loading on iOS (thanks @velebit!) ([\#14](https://github.com/zlovatt/obsidian-trim-whitespace/issues/14))
+- Typo in settings ([\#16](https://github.com/zlovatt/obsidian-trim-whitespace/issues/16))
+- Issue trimming code blocks containing template literals ([\#17](https://github.com/zlovatt/obsidian-trim-whitespace/issues/17))
+
+---
+
 ## [v0.2.2] - 2022/08/10
 
 ### Changed
@@ -67,7 +91,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 - Initial commit!
 
-[Unreleased]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/master...develop
+[Unreleased]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/main...develop
+[v0.3.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.2.2...v0.3.0
 [v0.2.2]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/zlovatt/obsidian-trim-whitespace/compare/v0.1.0...v0.2.0